+++ aws-android-sdk-rekognition/src/main/java/com/amazonaws/services/rekognition/model/ServiceQuotaExceededException.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright 2010-2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License").
  * You may not use this file except in compliance with the License.
@@ -20,8 +20,9 @@ import com.amazonaws.AmazonServiceException;
 /**
  * <p/>
  * <p>
- * The size of the collection exceeds the allowed limit. For more information,
- * see Limits in Amazon Rekognition in the Amazon Rekognition Developer Guide.
+ * The size of the collection or tag list exceeds the allowed limit. For more
+ * information, see Limits in Amazon Rekognition in the Amazon Rekognition
+ * Developer Guide.
  * </p>
  */
 public class ServiceQuotaExceededException extends AmazonServiceException {